# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Midanic E-Commerce & ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /erp/stores:
    get:
      operationId: getErpStores
      tags: [erp]
      summary: List all stores (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: createErpStore
      tags: [erp]
      summary: Create a new store
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /erp/stores/{id}:
    patch:
      operationId: updateErpStore
      tags: [erp]
      summary: Update a store (rename / activate / deactivate)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoreRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
    delete:
      operationId: deleteErpStore
      tags: [erp]
      summary: Delete a store (only if it has no data)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Products ──────────────────────────────────────────────────────
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: categoryId
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /products/{id}/reviews:
    post:
      operationId: createReview
      tags: [products]
      summary: Add product review
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ── Categories ────────────────────────────────────────────────────
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update category (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Cart ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    # Auth
...
    Store:
      type: object
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        slug:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, nameAr, nameEn, slug]
...
    CreateStoreRequest:
      type: object
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        slug:   { type: string }
        isActive: { type: boolean }
      required: [nameAr, nameEn, slug]
    UpdateStoreRequest:
      type: object
      properties:
        nameAr:   { type: string }
        nameEn:   { type: string }
        isActive: { type: boolean }
...
    Product:
      type: object
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        descriptionAr:
          type: string
        descriptionEn:
          type: string
        price:
          type: string
        imageUrl:
          type: string
          nullable: true
        stock:
          type: integer
        categoryId:
          type: integer
          nullable: true
        rating:
          type: string
        reviewCount:
          type: integer
        createdAt:
          type: string
      required: [id, nameAr, nameEn, price, stock, rating, reviewCount]
    Review:
      type: object
      properties:
        id:
          type: integer
        rating:
          type: integer
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
        userName:
          type: string
          nullable: true
      required: [id, rating]
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
    ProductsResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [products, total, page, limit]
    CreateProductRequest:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        descriptionAr:
          type: string
        descriptionEn:
          type: string
        price:
          type: string
        imageUrl:
          type: string
          nullable: true
        stock:
          type: integer
        categoryId:
          type: integer
          nullable: true
      required: [nameAr, nameEn, price]
    CreateReviewRequest:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required: [rating]

    # Categories
    Category:
      type: object
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, nameAr, nameEn]
    CreateCategoryRequest:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        imageUrl:
          type: string
          nullable: true
      required: [nameAr, nameEn]

    # Cart
...